Refactor numpy code, allow other subclassing for other libraries
remove whitespace

numpy white
