+++ sympy/printing/lambdarepr.py
@@ -1,8 +1,8 @@
 from .pycode import (
     PythonCodePrinter,
     MpmathPrinter,  # MpmathPrinter is imported for backward compatibility
-    NumPyPrinter  # NumPyPrinter is imported for backward compatibility
 )
+from .numpy import NumPyPrinter  # NumPyPrinter is imported for backward compatibility
 from sympy.utilities import default_sort_key
